# backend/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "PwndocNG API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "/",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/images/{imageId}": {
      "get": {
        "tags": [
          "Image"
        ],
        "description": "",
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Image"
        ],
        "description": "",
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/images/": {
      "post": {
        "tags": [
          "Image"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "auditId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/images/download/{imageId}": {
      "get": {
        "tags": [
          "Image"
        ],
        "description": "",
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/api/templates": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "file": {
                  "example": "any"
                },
                "ext": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/templates/{templateId}": {
      "put": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "file": {
                  "example": "any"
                },
                "ext": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/templates/download/{templateId}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}